cleaned up and divided into header and source files
